requirements.txt
Django==2.2.4
django-ckeditor==5.8.0
django-js-asset==1.2.2
django-taggit==1.1.0
isort==4.3.21
mysql==0.0.2
mysqlclient==1.4.4
pytz==2019.2
sqlparse==0.3.0
